Refactoring main entry points.
# Changed
- Similar logic from entry points are now put into the common main file.
